bq-query-generator/requirements.txt
# Vector Database
pinecone>=3.0.0

# Embeddings
openai>=1.0.0

# LLM for Query Generation
anthropic>=0.18.0

# Configuration
python-dotenv>=1.0.0

# Markdown Processing
markdown>=3.5.0

# CLI Framework
click>=8.1.0

# Syntax Highlighting
pygments>=2.17.0

# Utilities
pyyaml>=6.0.0
